requirements.txt
pathlib
python-dotenv
pandas
requests
pathlib
os
re
bs4
random
json
time
unicodedata
datatime
